requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.3
djangorestframework==3.15.2
gunicorn==23.0.0
mysqlclient==2.2.6
packaging==24.2
pillow==11.0.0
python-decouple==3.8
qrcode==8.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
